refactor: split in composables
