+++ app/components/Analysis/FlagItem.vue
@@ -1,14 +1,13 @@
 <script setup lang="ts">
-import type {
-  FlagDataPoint,
-  GitHubEvent,
-  IdentifyResult,
-} from "@unveil/identity";
-import dayjs from "dayjs";
+import type { IdentifyResult } from "@unveil/identity";
 
 type Flag = IdentifyResult["flags"][number];
 
 const props = defineProps<{ flag: Flag }>();
+
+const { isExceeded, parseDataPoint, groupDataPoints } = useFlagDataPoints();
+const { getEventDescription, getEventUrl, formatEventTime, groupEvents } =
+  useGitHubEventDisplay();
 
 const isExpanded = ref(false);
 const areEventsExpanded = ref(false);
@@ -28,233 +27,6 @@ const visibleEvents = computed(() =>
     ? props.flag.events
     : props.flag.events.slice(0, EVENT_PREVIEW_COUNT),
 );
-
-function getDataPointIcon(label: string): string {
-  const l = label.toLowerCase();
-  if (l.includes("branch")) return "i-lucide:git-branch";
-  if (l.includes("fork")) return "i-lucide:git-fork";
-  if (l.includes("merge") || l.includes("merged")) return "i-lucide:git-merge";
-  if (l.includes("pr") || l.includes("pull request"))
-    return "i-lucide:git-pull-request";
-  if (l.includes("closed")) return "i-lucide:git-pull-request-closed";
-  if (l.includes("star") || l.includes("watch")) return "i-lucide:star";
-  if (l.includes("comment")) return "i-lucide:message-square";
-  if (l.includes("push")) return "i-lucide:upload";
-  if (l.includes("repo")) return "i-lucide:folder-git-2";
-  if (l.includes("bounty")) return "i-lucide:trophy";
-  if (l.includes("ratio") || l.includes("%")) return "i-lucide:percent";
-  if (l.includes("entropy") || l.includes("diversity") || l.includes("types"))
-    return "i-lucide:layers";
-  if (l.includes("age") || l.includes("date")) return "i-lucide:calendar";
-  if (
-    l.includes("hour") ||
-    l.includes("interval") ||
-    l.includes("gap") ||
-    l.includes("window") ||
-    l.includes("span") ||
-    l.includes("(s)") ||
-    l.includes("(min)")
-  )
-    return "i-lucide:timer";
-  if (l.includes("consecutive") || l.includes("rapid") || l.includes("burst"))
-    return "i-lucide:zap";
-  if (l.includes("event") || l.includes("activity")) return "i-lucide:activity";
-  return "i-lucide:bar-chart-2";
-}
-
-function isExceeded(point: FlagDataPoint): boolean {
-  if (point.threshold === undefined) return false;
-  const v = parseFloat(String(point.value).replace("%", ""));
-  const t = parseFloat(String(point.threshold).replace("%", ""));
-  return !isNaN(v) && !isNaN(t) && v > t;
-}
-
-function parseDataPoint(point: FlagDataPoint): {
-  label: string;
-  displayValue: string;
-  displayThreshold: string | undefined;
-} {
-  const unitMatch = point.label.match(/^(.*?)\s*\(([^)]+)\)$/);
-  if (!unitMatch) {
-    return {
-      label: point.label,
-      displayValue: String(point.value),
-      displayThreshold:
-        point.threshold !== undefined ? String(point.threshold) : undefined,
-    };
-  }
-  const [, cleanLabel, unit] = unitMatch as [string, string, string];
-  const withUnit = (val: FlagDataPoint["value"]) => {
-    if (typeof val === "boolean") return String(val);
-    const n = parseFloat(String(val));
-    return !isNaN(n) ? `${val}${unit}` : String(val);
-  };
-  return {
-    label: cleanLabel,
-    displayValue: withUnit(point.value),
-    displayThreshold:
-      point.threshold !== undefined ? withUnit(point.threshold) : undefined,
-  };
-}
-
-function groupDataPoints(
-  data: FlagDataPoint[],
-): { icon: string; points: FlagDataPoint[] }[] {
-  const groups: { icon: string; points: FlagDataPoint[] }[] = [];
-  for (const point of data) {
-    const icon = getDataPointIcon(point.label);
-    const last = groups[groups.length - 1];
-    if (last && last.icon === icon) {
-      last.points.push(point);
-    } else {
-      groups.push({ icon, points: [point] });
-    }
-  }
-  return groups;
-}
-
-function getEventIcon(event: GitHubEvent): string {
-  switch (event.type) {
-    case "PullRequestEvent":
-      return "i-lucide:git-pull-request";
-    case "PushEvent":
-      return "i-lucide:upload";
-    case "CreateEvent":
-      return "i-lucide:git-branch";
-    case "DeleteEvent":
-      return "i-lucide:trash-2";
-    case "ForkEvent":
-      return "i-lucide:git-fork";
-    case "WatchEvent":
-      return "i-lucide:star";
-    case "IssueCommentEvent":
-    case "PullRequestReviewCommentEvent":
-      return "i-lucide:message-square";
-    case "IssuesEvent":
-      return "i-lucide:circle-dot";
-    case "PullRequestReviewEvent":
-      return "i-lucide:eye";
-    case "ReleaseEvent":
-      return "i-lucide:tag";
-    case "CommitCommentEvent":
-      return "i-lucide:message-circle";
-    case "MemberEvent":
-      return "i-lucide:user-plus";
-    default:
-      return "i-lucide:activity";
-  }
-}
-
-function getEventDescription(event: GitHubEvent): string {
-  const payload = event.payload as Record<string, unknown> | undefined;
-  switch (event.type) {
-    case "PullRequestEvent": {
-      const action = payload?.action as string | undefined;
-      const pr = payload?.pull_request as { number?: number } | undefined;
-      return pr?.number
-        ? `PR #${pr.number} ${action ?? ""}`.trim()
-        : `Pull request ${action ?? ""}`.trim();
-    }
-    case "PushEvent": {
-      const commits =
-        (payload?.size as number) ??
-        (payload?.commits as unknown[])?.length ??
-        0;
-      const ref = (payload?.ref as string)?.replace("refs/heads/", "") ?? "";
-      return ref
-        ? `${commits} commit${commits !== 1 ? "s" : ""} → ${ref}`
-        : `${commits} commit${commits !== 1 ? "s" : ""}`;
-    }
-    case "CreateEvent": {
-      const refType = payload?.ref_type as string | undefined;
-      const ref = payload?.ref as string | undefined;
-      return ref ? `${refType} created: ${ref}` : `${refType ?? "ref"} created`;
-    }
-    case "DeleteEvent": {
-      const refType = payload?.ref_type as string | undefined;
-      const ref = payload?.ref as string | undefined;
-      return ref ? `${refType} deleted: ${ref}` : `${refType ?? "ref"} deleted`;
-    }
-    case "ForkEvent": {
-      const forkee = payload?.forkee as { full_name?: string } | undefined;
-      return forkee?.full_name ? `forked → ${forkee.full_name}` : "forked";
-    }
-    case "WatchEvent":
-      return "starred";
-    case "IssueCommentEvent": {
-      const issue = payload?.issue as { number?: number } | undefined;
-      return issue?.number ? `comment on #${issue.number}` : "issue comment";
-    }
-    case "IssuesEvent": {
-      const action = payload?.action as string | undefined;
-      const issue = payload?.issue as { number?: number } | undefined;
-      return issue?.number
-        ? `issue #${issue.number} ${action ?? ""}`.trim()
-        : `issue ${action ?? ""}`.trim();
-    }
-    case "PullRequestReviewEvent": {
-      const pr = payload?.pull_request as { number?: number } | undefined;
-      return pr?.number ? `reviewed PR #${pr.number}` : "PR review";
-    }
-    case "PullRequestReviewCommentEvent": {
-      const pr = payload?.pull_request as { number?: number } | undefined;
-      return pr?.number ? `comment on PR #${pr.number}` : "PR comment";
-    }
-    case "ReleaseEvent": {
-      const release = payload?.release as { tag_name?: string } | undefined;
-      return release?.tag_name ? `released ${release.tag_name}` : "release";
-    }
-    default:
-      return event.type?.replace("Event", "") ?? "event";
-  }
-}
-
-function formatEventTime(date: string | null | undefined): string {
-  if (!date) return "";
-  return dayjs(date).format("MMM D, h:mma");
-}
-
-function groupEvents(
-  events: GitHubEvent[],
-): { icon: string; events: GitHubEvent[] }[] {
-  const groups: { icon: string; events: GitHubEvent[] }[] = [];
-  for (const ev of events) {
-    const icon = getEventIcon(ev);
-    const last = groups[groups.length - 1];
-    if (last && last.icon === icon) {
-      last.events.push(ev);
-    } else {
-      groups.push({ icon, events: [ev] });
-    }
-  }
-  return groups;
-}
-
-const ZERO_SHA = "0000000000000000000000000000000000000000";
-
-function getEventUrl(event: GitHubEvent): string | undefined {
-  const repo = event.repo?.name;
-  const payload = event.payload as Record<string, unknown> | undefined;
-  if (!repo) return undefined;
-  const base = `https://github.com/${repo}`;
-  switch (event.type) {
-    case "PullRequestEvent":
-    case "PullRequestReviewEvent":
-    case "PullRequestReviewCommentEvent": {
-      const pr = payload?.pull_request as { number?: number } | undefined;
-      return pr?.number ? `${base}/pull/${pr.number}` : undefined;
-    }
-    case "PushEvent": {
-      const after = payload?.after as string | undefined;
-      if (after && after !== ZERO_SHA) {
-        return `${base}/commit/${after}`;
-      }
-      return undefined;
-    }
-    default:
-      return undefined;
-  }
-}
 </script>
 
 <template>
@@ -275,8 +47,8 @@ function getEventUrl(event: GitHubEvent): string | undefined {
         class="mt-3 pt-3 border-t border-gh-border-light/30 space-y-2"
       >
         <div
-          v-for="group in groupDataPoints(flag.data)"
-          :key="group.icon"
+          v-for="(group, i) in groupDataPoints(flag.data)"
+          :key="`${group.icon}-${i}`"
           class="flex gap-2"
         >
           <span
@@ -289,9 +61,7 @@ function getEventUrl(event: GitHubEvent): string | undefined {
               :key="point.label"
               class="flex flex-col gap-0.5"
             >
-              <span class="text-gh-muted text-m">{{
-                parseDataPoint(point).label
-              }}</span>
+              <span class="text-sm">{{ parseDataPoint(point).label }}</span>
               <div class="flex items-center gap-1">
                 <template v-if="typeof point.value === 'boolean'">
                   <span
@@ -303,13 +73,7 @@ function getEventUrl(event: GitHubEvent): string | undefined {
                     class="text-xs"
                   />
                 </template>
-                <span
-                  v-else
-                  class="text-sm"
-                  :class="
-                    isExceeded(point) ? 'text-gh-danger-hover' : 'text-gh-text'
-                  "
-                >
+                <span v-else class="text-sm text-gh-muted">
                   {{ parseDataPoint(point).displayValue }}
                 </span>
                 <span
@@ -330,8 +94,8 @@ function getEventUrl(event: GitHubEvent): string | undefined {
             Evidence
           </p>
           <div
-            v-for="group in groupEvents(visibleEvents)"
-            :key="group.icon"
+            v-for="(group, i) in groupEvents(visibleEvents)"
+            :key="`${group.icon}-${i}`"
             class="flex gap-2"
           >
             <span
